Fixed some animation timing and rounding issues
This made some animations better, some worse but it's more technically correct
and closer to a working solution

+++ lotte/src/main/java/com/airbnb/lotte/animation/LotteAnimatableColorValue.java
@@ -88,7 +88,7 @@ public class LotteAnimatableColorValue implements LotteAnimatableValue<Integer> 
             for (int i = 0; i < keyframes.length(); i++) {
                 JSONObject keyframe = keyframes.getJSONObject(i);
                 long frame = keyframe.getLong("t");
-                float timePercentage = (frame - startFrame) / durationFrames;
+                float timePercentage = (float) (frame - startFrame) / (float) durationFrames;
 
                 if (outColor != null) {
                     colorKeyframes.add(outColor);
